// src/types/bindings/CreateBusinessProfileInput.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BankAccount } from "./BankAccount";
import type { EntityType } from "./EntityType";

export type CreateBusinessProfileInput = { entity_type: EntityType, name: string, address: string | null, email: string | null, phone: string | null, ssm_no: string | null, nric: string | null, sst_no: string | null, bank_accounts: Array<BankAccount>, enabled_payment_methods: Array<string>, default_tax_rate: number | null, default_quotation_valid_days: number, default_invoice_due_days: number, data_dir: string, };